started plotting yearly energy

+++ Mothercell/Hive.IO/Hive.IO/Plots/YearlyEmissionsPlot.cs
@@ -1,12 +1,12 @@
 ﻿namespace Hive.IO.Plots
 {
-    public class YearlyEmissionsPlot: AmrPlotBase
+    public class YearlyEmissionsPlot : AmrPlotBase
     {
         protected override string Title => "CO₂ Emissions";
         protected override string Unit => "kgCO₂";
-        protected override double EmbodiedSystems => Results.EmbodiedEmissionsSystems;
-        protected override double EmbodiedBuildings => Results.EmbodiedEmissionsBuildings;
-        protected override double OperationSystems => Results.OperationEmissionsSystems;
-        protected override double OperationBuildings => Results.OperationEmissionsBuildings;
+        protected override float EmbodiedSystems => (float)Results.EmbodiedEmissionsSystems;
+        protected override float EmbodiedBuildings => (float)Results.EmbodiedEmissionsBuildings;
+        protected override float OperationSystems => (float)Results.OperationEmissionsSystems;
+        protected override float OperationBuildings => (float)Results.OperationEmissionsBuildings;
     }
 }
